requirements.txt
django==1.10.5
djangorestframework==3.5.4
python-decouple==3.0
